requirements.txt
Flask==3.0.3
gunicorn==21.2.0
requests==2.31.0
hangul-romanize==0.1.0supabase
python-dotenv
